refactor(tests): split mx checker suite by scenario
Replace the monolithic MX checker test module with required, optional, and error-focused suites, add shared MX test support helpers, and strengthen negative-path assertions with stable message checks while preserving full coverage.
